refactor: simplify CustomCancellationToken
