+++ src/client/llmClient.ts
@@ -1,8 +1,10 @@
-import type { CustomCancellationToken } from "@/tasks";
-
 export interface LLMClient {
     generatePageContents(token: CustomCancellationToken, prompt: string, model: string | null, locale: string | null): Promise<string | null>;
     decorateContents(token: CustomCancellationToken, prompt: string): Promise<string | null>;
+}
+
+export interface CustomCancellationToken {
+    onCancellationRequested(listener: () => void): any;
 }
 
 export class UnconfiguredClient implements LLMClient {
